+++ src/api/contract/approve.js
@@ -22,7 +22,7 @@ const validateDepositAmountAndConvertToBN = amount => {
 
 const tokensWhichNeedResetToZero = ['USDT', 'OMG']
 
-module.exports = async (dvf, token, deposit, spender = dvf.config.DVF.starkExContractAddress, chain = 'ETHEREUM') => {
+module.exports = async (dvf, token, deposit, spender = dvf.config.DVF.starkExContractAddress, chain = 'ETHEREUM', options = {}) => {
   if (token === 'ETH') {
     // TODO: This code is not very safe if caller expects the result to be of
     // the shape returned by dvf.eth.send below.
@@ -52,7 +52,9 @@ module.exports = async (dvf, token, deposit, spender = dvf.config.DVF.starkExCon
     [
       spender,
       amount
-    ]
+    ],
+    null,
+    options
   )
 
   // For some tokens, the amount needs to be reset to 0 before setting it to